Added Gordos's new design of icons and splash screen.
The source Photoshop files for the new design are
stored in the new directory 'psd'.

In the icons directory, the splash screen is now
common for all themes and stored icons/about_wings.bmp.

In wings_io, there is now support for handling icons that
don't have power-of-two sizes to handle the new splash screen.

For Mac OS X, the old document icon for Wings files has been updated
to contain the new icon. The document icons for other file formats
that Wings could handle (such as 3DS) have been removed. The plist
file and the Xcode project file has been updated to reflect those
changes.