Updated the README, BUILD.*, and AUTHORS files.
The main change in BUILD.unix and BUILD.win32 is that
R13B is now required. Also changed the version numbers
in all examples to wings-1.0.

One important developer was completely missing from the
AUTHORS files. Also put all the credits for the translations
in a separate section and added all translators that I could
find in the NOTES file.