The new icon was not used in all places.
In Windows, the original Wings icon was shown at the top left of
the Wings window, and also when switching applications
using Alt+Tab. I have replaced it with the new icon, but I could
not use any transparancy due to bugs in ESDL and SDL. (Furthermore,
SDL only supports hard-edged mask, which would probably not have
looked good.)

On the Mac, the icon that is shown in the Dock and in the task switcher
(Cmd+Tab) is now an 128x128 icon (without transparancy,
unfortunately).