Corrected a problem with hard edges flickering.
I made sure that hard_edge_width/1 would never return anything
smaller than 1. It used to return 0 if the hard edge width was
set to 1 and the selection mode was not edge. Since 0 is an
illegal line width, hard edges would be drawn with the same
width as the highlighted or selected edges.