Now using a mask for the Wings icon.
A hard-edge mask is now used for the icon set by sdl_video:wm_setIcon/2.
The mask for the icon is in the WBM format, which is a very simple
bitmap format (trivial to load) and can be exported from Photoshop.

I have also added a README file to the psd directory to explain
the purpose of each file.