Eliminated incorrect use of array:sparse_size/1.
array:sparse_size/1 does NOT calculate the number of
non-default entries in an array, as I initially thought/hoped.

Therfore, I have added a new function wings_util:array_entries/1 to
calculate the number of non-default entries for an array. Since
this function must go through all entries in an array, it is slower
than a simple gb_trees:size/1.

In wings_subdiv, gb_trees:size(Etab) (and array:sparse_size(Etab),
incorrectly) was used to find whether we were operating on the
entire object or only a part. I turns out that the call to
array:sparse_size/1 could be easily eliminated, as we already
knew from the entry point (smooth/1 or smooth/5) whether
we were operating on the entire table.

In wings_draw, gb_trees:size(Etab) was compared to the number of
selected edges to choose the most optimal way to draw the selected
edges. I removed the test and kept the more general drawing code
that works on any number of selected edges.