Clean-up of Windows icon.
I was noticing some color banding in the task switcher icon in
Windows, and found out that converting the BMP to 256 colors seemed to
fix it. While I was at it, I also cleaned up the edges a little bit on
both that BMP file and also the WBM used for the cutout. I was just
editing the files that came with the install of Wings RC3, so new
versions of those are attached.

[I updated the wings_icon_small.psd file to match those changes. /Bjorn]