Operate directly on the edge table.
When the edge table was a gb_tree, massive updates to the edge table
was better done by first converting the gb_tree to an orddict, work
on it, and convert back.

For an array, it is better to operate directly on the array to
increase sharing between undo states.

The speed is about the same, perhaps slightly slower. I choose to do
this change because the code is simpler and not converting the edge
array to an orddict and back will increase sharing between undo states
when smoothing part of the object in face mode (in body mode, all of
the edge table will be updated anyway).