Implement a Develop menu.
It is becoming increasingly tedious to insert and remove code for
measurements and printing out internal data structures. Therefore I
have implemented a Develop menu which can be turned on in the Misc
tab of the preferences (by default it is off).

The actual menu is empty in this initial commit.