Add more types in records and eliminate Dialyzer warnings.
Notes:

AutoUV uses #st.bb and #we.name for its own purposes,
so types must allow for that. (For now.)

erlang:make_tuple/3 is not known by Dialyzer in R13B.
It turns out that the code in wings_undo that used
make_tuple/3 can be made shorter and more maintainable
by using record_info/2.

The field #we.saved is re-used for storing an image index
in an ugly way.