Not rebuilding the edge table when extruding single faces.
Here is another case when there was previous optimization
of not updating the edge table directly, but instead collecting
new edges into a list, and rebuilding the gb_tree for the edge
table at the end.

Directly updating the edge array is simpler, slightly faster,
and requires less memory in the undo queue (especially when
extruding only a few faces from a huge object).