Fix crash in Aim when targeting unselected geometry.
Bjorn's change to save the last command executed in #st.last_cmd
assumed that command/2 always was called with a #st{} record
as the second argument. That was not so. Aiming at unselected
geometry (if turned on in the preferences), would construct
a second argument looking like {add,St0,St} or {delete,St0,St}.

I have changed that to send the extra information in the command
itself. There is also no need to send the entire St, since only
the selection is used.