Optimized drawing by by using a single interleaved array
I saved some time by putting vertex positions in the same
array. This could be extended to always put all data in one
array, and we could save information about the layout of the
array (stride and so on) so that we could set up all arrays
properly.

The stride is stored with each buffer pointer, so that code
outside of wings_draw_setup, will not need to know
(or have any need to know) whether a given buffer
is interleaved or not.