Added Plug-in for Generating Ambient-Occlusion via OpenGL.
Uses hardware-accelerated OpenGL calls, instead of ray-tracing, to
calculate and set a per-vertex ambient-occlusion factor. A HemiCube is
used to sample the environment's visibility. The results are stored in
the vertex-colors and can also be baked to a texture through autouv.

The Hemires variable is hard-coded at 64. Values greater than 64
slowed down the calculation, and made no difference in the result.
Values below 64 offered slightly increased speed at the cost of a
much worse-looking solution. 64 seems to be the magic number here.