Handle degenerate faces when showing smooth normals
Degenerate faces that have lost of one or more triangles in the triangulation
would cause a badmatch (because if we see two triangles after triangulation,
we expect the original face to have exactly four vertices). Avoid the
crash by using the face normal for all vertices if the original face
doesn't have the expected number of vertices.