Protect vertex buffers from garbage collection
If the last use of a binary is in a call to gl:vertexPointer/4,
the binary could very well be deallocated before it is used by
(for example) gl:drawArrays/3. Make sure that the binary is
kept by passing it to a dummy function after the call
to gl:drawArrays/3.

You might think that vertex arrays kept in #dlo{} records are
protected from garbage collections, but it turns out that
temporary #dlo{} records are created in several places.
Therefore, draw_faces_all/2 and smooth_faces_all/2 also need
to protect their D (#dlo{}) arguments.