Do not crash on stale hotkeys
Hotkeys could have been created in ancient version of Wings or
in a plug-in that is now disabled. It is annyoying to get a
crash dump if you use a stale hotkey.

Therefore, if a command that is initiatied by a hotkey fails
in any way, show a dialog box asking the user what to do.
In the dialog box we allow the user to ignore the error or to
delete the offending hotkey. (We do not give a "Crash Dump"
dump button, because it seems strange to generate a crash
dump later than the actual error occurred.)

To handle the related problem that repeating a command may
cause a crash because a plug-in has been disabled, have the
plug-in manager delete the last repeatable command (so that
it cannot be repeated) if any plug-in was enabled or disabled.

Handle stale hotkeys in Tweak mode in the same way.

AutoUV does not yet handle stale hotkeys.