Only use multisampling with OpenGL implementations that support it
Test for the GL_ARB_multisample extension and only enable the
choice for multisampling in the preferences if the OpenGL implementation
supports it. That will avoid (harmless) error messages to the console
if Develop|OpenGl Errors has been enabled.

Also remove defines such as SDL_GL_MULTISAMPLEBUFFERS from
wings_init that there only there to support an old ESDL (that
will not work anyway for other reasons).

Update Help|OpenGL Info to report the number of sample buffers.