Add a script to extract release notes from check-in comments
Since we started to have separate branches for every
change, there will typically be a merge conflict
on NOTES-1.1 every time a branch is merged into 'pu'
or 'next'.

To avoid the merge conflict, we will put the release
notes into the check-in comments. The next paragraph
is a release note.

NOTE: For developers: Release notes to be included in a release
should no longer be directly added to NOTES-1.1 files, but
instead be written into the check-in comment. Everything in
a comment from "NOTE:" (without the quotations marks) at
the beginning of a line to the end of the check-in comment,
will be included in the release notes.

To extract the release notes, run the script tools/extract_notes.
It will extract and print all notes for all commits after
the tag for the latest release. [bjorng]