Implement a driver for picking instead of using OpenGL
Since very few applications use picking/selection support
in OpenGL, it tends to be poorly supported in many OpenGL
drivers. Therefore, implement a driver to do picking
ourselves. An advantage of doing it this way is
that the pick code in the driver can directly use the
vertex array that is used for drawing.

NOTE: Selection (finding out which face is under the mouse
cursor) used to be done using a special selection render
mode in OpenGL. However, this mode does not work well
in all OpenGL drivers. Therefore I have implemented my
own code for selection handling in C as a driver for the
Erlang emulator. (Thanks to FonteBoa for testing.) [bjorng]