Make wings_face:fold/4,5 vertex order consistent with other fold/info functions
Compared to other fold and info functions (for instance,
wings_face:fold_vinfo/4,5), the start vertex is different.
For instance, if fold/4 gives the sequence 0, 1, 2, 3,
then fold_vinfo/4 gives the 1, 2, 3, 0, which will result
in wrong results if fold/4 is used to retrieve vertex
positions and fold_vinfo/4 is used to retrieve UV
coordinates or vertex colors.

Change fold/4,5 to be consistent with the other functions.
The rule is that the vertex nearest to the successor
pointer should always be presented first (i.e. on the
left side, use the "vs" vertex since it is nearest to the
the "ltsu" pointer).

This change seems to be safe after a quick look at the
code.