Speed up Select|Similar when the seed selection contains similar elements
After having calculated the templates for all elements
in the selection, we do an ordsets:from_list/1 to eliminate
duplicates. The problem is that two templates that are
ALMOST equal will both be kept in the list.

To get rid of a template that is almost equal to another
template, first sort the list of templates, then go through
the list and match templates against each other using
match_template/2. If they match (i.e. if they are similar),
only keep the first template. That will leave us with a
minimum number of templates that should match the same
elements.

Also change the comparison function to the one described
by Christer Ericson in Real-Time Collision Detection, and
simplify the creation of face templates by using
wings_face:vertex_positions/2.

NOTE: Select|Similar is now faster if the initial selection
contains elements that are similar to each other.
(Thanks to Ran13.) [bjorng]