Eliminate references to #edge.a and #edge.b
The structure of the original code makes it difficult
to use the wings_va module to update vertex
attributes.

Fortunately, there is a very simple solution:

Welding vertex A to vertex B is almost exactly the
same operation as collapsing the edge between A and B.
The only difference is that vertex B should keep
its position.

Therefore, this module can be dramatically simplified
by using wings_collaps:collapse_edge/3 and moving
back the surviving vertex to its original position.