Reimplement wings_va with new data structures
Instead of storing vertex attributes in the edge table,
we have two fields #we.lv and #we.rv. The 'lv' field
(unless it is 'none') contains an array indexed with
edge number, storing attributes on the left side of the
edges (for the 'vs' vertex). Similarly, the 'rv' array
stores all attributes on the right side of edges (for
the 've' vertex).