Don't lose UV coordinates for edges that only have UV coordinates on one side
Also make sure that wings_va:vtx_attrs/3 returns 'none' and
not 'not_done' if called with invalid vertex and/or face
arguments. A call with invalid arguments can happen if hole
faces automatically created by imported are tesselated, because
the hole-filling code in wings_we may construct faces that
share vertices in such a way that a vertex can be reached
from a face, but that face cannot be reached from the vertex.

NOTE: Re-exporting an imported model with tesselation could
cause a crash. Also, when importing objects that had UV coordinates
on some faces but not all, UV coordinates would be discarded
on both side of edges that only had UV coordinates on one side.
(Thanks to Siber.) [bjorng]