Eliminate selection glitches
In rare circumstances, it is possible to "select through" a face and
select elements behind it. This glitch occurs because the wpc_pick
module does not provide any hints about how far away each face is
from the eye point (i.e. the near clipping plane). Instead, code in
wings_pick attempts to find the "best face", not using a proper
intersection test, but through a heuristic method that seems to work
*most* of the time. (Basically, for each face, the distance to the
plane of the face is calculated and weighed by the inverse of the
the dot product of its normal with they eye point ray; the shortest
distance win.)

Eliminate these selection glitches by letting wpc_pick calculate
depth values for the clipped triangle vertices (picking the face
with the *lowest* depth value which is nearest the eye point).

NOTE: Eliminate selection glitches where elements behind a face
could be selected through the face. (Thanks to deerwood.) [bjorng]