Eliminate crash in the progress bar process
Also use timer:now_diff/2 instead of our own now_diff/2
(timer:now_diff/2 was not exported when this module was
first implemented, but now it is).

NOTE: In rare circumstances, the progress bar could reach 1.0
(the far end) even if the operation was not finished and cause
Wings to crash hard. (Thanks to deerwood.) [bjorng]