Import: Remove shared vertices by duplicating them
Wings' modeling commands does not allowing creation of
non-manifold geometry. However, when importing a model,
it is possible to create shared vertices (a vertex that
is shared by geometry that does not share an edge).

Rewrite wings_we:build/2 to eliminate shared vertices
by duplicating them.

Also change wings_we:build/2 to return the atom 'error'
when it encounters an error, instead of throwing an exception,
and change the import to stop catching errors because it
can hide bugs.

NOTE: Shared vertices are now eliminated on import. (The
modeling commands in Wings does not allow creating of shared
vertices (also called waists), i.e. faces that only share vertices
but no edges, so it should not be allowed on import.) [bjorng]