Hide the virtual mirror seam
Calculate vertex normals (used in the non-work mode display)
as if the mirror has been frozen.

Most of the implementation is straightforward (and explained
in comments). But there is one detail worth mentioning:

We want to keep wings_we:normals() functional (i.e. the result
should only depend on the arguments, not on other state such
as the process dictionary). The problem is that we need the
mirror matrix. We could calculate it from the scratch,
but that seems wasteful since we know that it has already been
calculated. Therefore, add another argument to wings_we:normals/2,
making it wings_we:normals/3, where the new argument is the
mirror matrix (or 'none').

NOTE: When displaying objects that have a virtual mirror face in
the smooth preview mode (i.e. turning off workmode by pressing TAB),
the seam between the real part of the object and mirror image will
now be soft. The Smooth command (and the Smooth Proxy mode) will
still sub-divide faces around the virtual mirror face as if the
edges were hard, so Smoothing before freezing the mirror and
after freezing may still not give exactly the same result
(how close the result is depends on the object; for some objects
the difference may not be noticeable). [bjorng]