Smoothed Preview: Prevent crash if there is an area light in the scene
Lights are not supposed to be proxyable, but there are only
tests at some entry points to prevent lights from being proxied.
Add the missing test to setup_all/2.

NOTE: Switching to the Smoothed Preview mode would cause a crash
if there were any area lights in the scene. (Thanks to Micheus.)
[bjorng]