AutoUV: Take advantage of co-existence of vertex colors/UV coordinates
NOTE: AutoUV used to hold a copy of the scene to be able to
generate a texture with vertex colors or materials.

This extra scene copy has been eliminated. When doing the initial
UV mapping of an object without previous UV coordinates, the
materials on the object will be converted to vertex colors (this
will only be done if there are no previous vertex colors in the
object AND if the object has non-default materials). The vertex
colors in the object can later be used to generate a texture.

With this change, vertex colors generated by the Ambient Occlusion
command can now be baked into a texture. [bjorng]