Fix problems with the View|Show commands
NOTE: Using one of the commands Show Colors, Show Materials, or
Show Textures in the View|Show sub-menu, the scene would not be
immediately updated to reflect the change. Furthermore, if
Show Colors was turned off and there was an object with vertex
colors in the scene, there would be a crash. [bjorng]