Eliminate the object mode (vertex color/material)
NOTE: The object mode (vertex color/material) for objects has
been eliminated.

Instead materials have a Vertex Colors setting that can be
"Ignore" (don't use vertex colors; default for new materials),
"Set" (to use vertex colors instead of the colors in the material),
and "Multiply" (to multiply vertex colors with the diffuse texture).

The material "default" has "Set" as the value for the "Vertex Colors"
property, meaning that if you assign vertex colors to face having
the "default" material, the vertex colors will be visible.

There is a new command (with a sub-menu) in the body mode menu called
"Vertex Attributes".

When opening a .wings file written by Wings 1.1.7 or earlier, the
object modes for the objects in the file will be ignored. If there
are vertex colors, they will be visible on faces that have
the 'default' material assigned. For faces with other materials,
the colors or textures in the material will be shown (even if
the object was in vertex color mode).

When saving a .wings file in this version of Wings and opening
it in an older version of Wings, all objects will be in
material mode (even if they have vertex colors).

For plug-in writers: #we.mode has been removed. It also means that
plug-ins that use #we{} needs to be re-compiled. [bjorng]