Unify handling of vertex colors and materials in display code
When an object is in vertex color mode, materials are completely
ignored. As we soon will remove the modes and let the materials
decide whether vertex colors should be displayed, we must always
use materials.

Therefore, rewrite the display for vertex color mode to be use
the same internal representation as plain objects and objects
with UV coordinates. That will actually simplify the display
code.

As part of this change, define a new API in wings_draw_setup for
both setting up and activate an array (for vertex, normal, color,
and texture coordinate arrays), and also functions for disabling
an array.

This new API simplifies the drawing code and is safer, as there is no
risk that we will accidentally enable an array that does not exist
(we can remove a boolean that was used to indicate whether a
material had an associated texture coordinate array).