Set correct default for array in wings_va:gc/1
wings_va:gc/1 rebuilds the arrays used to hold vertex attributes,
but does not set the default value of the new arrays to 'none'
as it should, but to 'undefined'. Having 'undefined' in the array
will cause crashes later unless there are defined vertex attributes
for all vertices in the object.