AutoUV: Optimize wings_va:set_vtx_face_uvs/4
When something has been updated in a chart, all UV coordinates in
the chart will be updated by calling wings_va:set_vtx_face_uvs/4.
If, for instance, only a face has been moved, most of the UV
coordinates in the chart have not actually changed.

Make sure that wings_va:set_vtx_face_uvs/4 does not update the
array with the vertex attributes unless the UV coordinates
have truly changed. That will generate less garbage and reduce
the size of the Undo queue.

We can also rearrange the tests so that we avoid tests that
can't possible succeed. That change reduces the typical time for
one call from about 11 microseconds to about 7 microsecond on
my computer.