Fix problem with wings_subdiv:inc_smooth/2 not updating vertex attributes
Mend wings_subdiv:inc_smooth/2 so that it will update vertex attributes
if they have been changed. For the moment, do it by re-smoothing the
model.

Also clean up the return value for wings_proxy:proxy_smooth_1/3.
It used to return a tuple with either 'false', 'inc', or 'smooth'
as the first element. Since the difference between 'inc' or 'smooth'
was not used by the caller, and may no longer be accurate since
wings_subdiv:inc_smooth/2 may force a full smooth, return 'true'
in both cases.

NOTE: In the Smoothed Preview mode, changes to vertex colors or
UV coordinates would not be visible. (Thanks to FonteBoa.) [bjorng]