Fix glitches in View|Show {Colors,Material} commands
Historically, there has been many display glitches in the
View|Show * commands; that is, that the display was not
immediately updated after toggling on of those commands.

Eliminate one reason for the glitches by postponing display
decisions. Currently, wings_draw_setup:prepare/3 will group
faces according to materials differently depending on the
settings of Show Materials and Show Colors. (Putting all
faces in the 'default' material group if Show Colors is off
is now clearly wrong, as that will disable display of materials
as well as of colors.)

Have wings_draw_setup:prepare/3 always produce the same face
groups regardless of the settings. Instead, have the display
code change the material for all faces to 'default' if
Show Materials is off. (That means that we will only have to
invalidate the display lists, not the vertex buffers, when
toggling the Show Materials setting.)

wings_draw_setup:prepare/3 will still test Show Colors and Show
Textures to not waste time and memory including vertex colors and/or
UV coordinates in the vertex buffers if they will not be used.