Fix display glitches in Smoothed Preview mode
When using the View|Show Materials and View|Show Textures, the
display will not always be immediately updated to reflect the
change. The problem is that wings_proxy:invalidate_dl/2 only
invalidates display lists, not vertex buffers.

To fix this problem, rename the function to invalidate/2
(since it now can invalidate more things than just display lists)
and change the second argument (what to invalidate) to make more
sense: 'vab' for vertex buffers and display lists,
'dl' (was 'all') for only invalidating display lists. Remove the
'material' value (it does the same thing as the old value 'all').
Keep the other values. And, of course, update all callers to pass
in the correct value.

NOTE: Fixed the bugs in the View|Show Colors and View|Show Materials
reported by Micheus. Also made sure that a change to the vertex
color option of a material would immediately be seen in the
scene. [bjorng]