Draw undefined characters as filled boxes instead of crashing
If the preferences text_display_lists is false (i.e. not using
display lists for characters below 256), there will be a crash
when attempting to draw undefined characters (for instance,
characters in the range 127-159).

As undefined characters may occur if the encoding of filenames
is not understood, we always want to fall back on drawing a
filled box instead of crashing when ask to draw undefined
characters.

NOTE: If the workaround against disappearing text was enabled,
there would be crash when attempting to display any character
not included in the fonts, which could happen when using the
file dialog on Linux to browse a file system where the filenames
are encoded using UTF-8. (Thanks to mifth.) [bjorng]