Mac OS X file dialog: Prevent crash when UTF8 filenames are chosen
The file dialog plug-in for Mac OS X will crash if a filename is
chosen that cannot be represented in 7-bit ASCII (for instance,
a filename containing the Swedish letters ÅÄÖ). Fix this by
having the driver return the filename encoded in UTF-8.

NOTE: In the file dialog on Mac OS X, selecting a filename
containing non-US ASCII characters (such as the Swedish
letters Å, Ä, or Ö) would cause a hard crash. [bjorng]