Prettify UTF-8 encoded filenames in the File menu
We continue to keep filenames in their neative encoding as
they are returned from the file dialog plug-ins, but for
display purposes in the File menu, if a filename *seems* to be
encoded in UTF-8, we will convert the filename to Unicode
characters and display it as such. It is possible that a
filename encoded in ISO latin1 could also look like a valid
UTF-8 filename, but it is not very likely in practice.
(And if the filename is encoded in some other format, we
have no way of displaying properly anyway, so it doesn't
matter if we happens to think that it is UTF-8.)

(This change should have been included in 1.1.9, but was
forgotten because I forgot it when cherry-picking commits
to place the first two commits in the on the maint branch
and the remaining on the next branch.)

NOTE: File dialogs would crash on Linux in the 1.1.9
release. [bjorn]