Fix distorted UV coordinates for edge extrude
Vertex attributes (UV coordinates and vertex colors) will be
incorrectly interpolated when edges are extruded in edge mode.
The reason is that wings_edge:cut/3 is used to cut an edge in the
middle, and the new vertex is then moved to the correct position.
Correct this problem by calculating the new position before
cutting and cut using wings_edge:fast_cut/3 to ensure that
the vertex attributes are properly interpolated.

NOTE: Extrude in edge mode would distort UV coordinates and
vertex colors. That has been corrected. [bjorng]