Fix broken Tools|Snap Image
Commit ddad5076ed6a36b6989d57e9b73681a2cc2269d6 broke
Tools|Snap Image so that it would crash most of the time.

Instead of fixing only that problem, rewrite the updating
of UV coordinates to follow the same approach as in AutoUV,
which is faster and preserves vertex colors.

Also make sure that vertex colors (if any) do not override
the texture. If the vertex color attribute for the material
is 'Set', change it to 'Multiply' so that vertex colors and
the texture will be combined.

NOTE: Tools|Snap Image was broken in several ways. It will now
also preserve vertex colors. (Thanks to PuzzledPaul.) [bjorng]