Fix saving/loading of area light geometry
Commit af2d517f9ee77a87f6f94b4ae82ee80d1ae2d542 changed the
definition of the e3d_face record, which broke compatibility
with .wings files containing area lights (since the geometry
for area lights are saved using the e3d_mesh and e3d_face
records).

Ideally, we should not use records at all for saving the
geometry of area lights, but changing the format requires
careful planning and implementation to preserve compatibility,
so we will not do that now.

Instead, change the saving of area lights so that they are
saved using the definition of e3d_face before 1.1.9. When
loading area lights, handle both the 1.1.9 e3d_face definition
and the classic definition. (We do all this by manipulating
records as tuples using element/2 and setelement/3.)

NOTE: Wings 1.1.9 (and 1.1.9.1) would crash when attempting to open
a .wings file containing area lights saved by a previous version of
Wings. This has been corrected.  Furthermore, when saving area
lights, they will be saved in the same way as in versions before
1.1.9. That means that a .wings file containing area lights saved
by this version of Wings cannot be opened by 1.1.9.x, but it can
be opened by older versions such as 1.0.1 and 1.1.8.
(Thanks to Micheus.) [bjorng]