Introduce the Face|Hole commands
Implement the Face|Hole|Create Hole to create one or more holes
from the selected faces, and Face|Hole|Remove Hole to remove
all holes adjacent to selected faces.

NOTE: Wings have had the _hole_ material for a long time to fake
holes by marking faces that should not be included when exporting.

This release introduce the command Face|Hole to create a holes
that looks like holes also in the Geometry window.

To remove a hole, select a face adjacent to the hole and RMB
click on the Face|Hole command to create a face to cover the
hole.

When loading old .wings files, all faces containing the _hole_
material will be converted to the new type of hole.

If you load a .wings file saved with this version of Wings in
an old release, the holes will become hidden faces, but they
will NOT have the _hole_ material (i.e. they will not behave
as holes when exported). [bjorng]