wings_facemat: Add is_material_used/2
Add is_material_used/2 to determine whether a specific material is
used in an object.

Also change to mat_faces/2 to not attempt to remove invisible faces
(I wrote "attempt", because it would crash if the input list contained
invisible faces and the object had more than one material).

Update the description for all/1 to note that only visible faces
are included in the result.