wings_draw: Fix crash when static faces lacks normal information
When splitting up display lists in a static part and a dynamic part
in preparation for a drag operation, it was implictly assumed that
any newly created face (thus lacking any entry in the normals array
with triangulation and surface normal) would be in the dynamic part.

However, with the introduction of holes, at least the Intrude
command may introduce new faces that appear in the static display
list (because the new faces do not share any vertices with any
face that are being moved).

Fix this by adding triangulation and normal for any face in the
static part to the normals array before drawing the faces.