Collapse vertices near holes in a smarter way
During the Collapse operation, when the creating the new edges
that will make up the face to replace the vertex being collapsed,
we do not care on which side of the edge we put the existing face
and where we put a new face. It does not really matter when both
faces are visible, but if we Collapse a vertex on the border of
a hole, the hole can disappear or be turned into a triangle if
we put it onto the wrong side of the edge.

Therefore, take care to keep existing faces "outside" and create
new faces "inside".