Fix display glitches by avoiding pushClientAttrib() and popClientAttrib()
When a material says that vertex colors should be ignored,
we use gl:disableClientState(?GL_COLOR_ARRAY) to suppress
display of vertex colors for the faces that follow. To re-enable
display of vertex colors for subsequent materials, we use
pushClientAttrib() and popClientAttrib() at the outer level.

Unfortunately, it seems that some OpenGL implementations lose
the stride settings for one or more vertex buffers when those
calls are used. (At least, that is how I interpret the interesting
display glitches that result.)

I have observed that problem on my Mac with OpenGL 2.0 running on
an Nvidia GeForce 8800 GT, but not on Windows XP running OpenGL 1.1
in software.

Avoid the problem by using gl:enableClientState(?GL_COLOR_ARRAY)
explicitly to re-enable the color buffer (but only if there is
a color buffer and it was disabled).