Clean up wings_edge_loop
The current code organization is just too confusing.

Move all exported function to the beginning of the file and
comment the purpose of each one. Add some block comments to
indicate which block of functions are helper for which exported
functions. Move a few common utility functions to the end of
the module. Shorten some overlong lines.