Simplify and comment Select|Edge Loop|Edge Loop
The existing code is too convoluted to allow modifications
(for example, to better support holes).

First, the deep list of components (called links in the original
code) and end points within the components, can be flattened
to a list of end points (with the component number included in
the end point tuple). That eliminates one function.

Second, use a queue to store the end points. That eliminates
one function argument.

Third, when it is not possible to expand further from an end point
because there is an odd number edges emanating from the end point
vertex, the selection generated up to that end point is not supposed
to be used if another end point in the same end point hit another
component. Keep track of those selections excplitly in a new argument
called "Stuck" (instead of the confusing implicit mechanism used in
the original code).