Snow Leopard: Force use of gcc 4.0
To still support building a MacOS 10.4 (Tiger) target on a
Snow Leopard development host, we must make sure that we use
the gcc-4.0 compiler and not the default gcc 4.2 compiler.
(I have no way to test Wings on either Tiger or PPC Macs.)

It is probably still possible to build on 10.5 (Leopard) host,
but I have not tested that.

NOTE: This release for Mac OS X has been built with Snow
Leopard (10.6) as the host OS. It is still supposed to work on
Tiger (10.4) with either PPC or Intel CPUs, but I have no way to
actually test that. [bjorng]