Save white vertex colors
Vertex colors that are white (i.e. (1.0,1.0,1.0)) are not saved,
making it look like that those vertices had no colors assigned when
loading the scene. In most circumstances, a missing color will default
to white, so that is usually not noticed. But it is a little bit ugly
that the distinction between a white vertex color and no vertex color
is lost when saving a scene, so we better fix it.

NOTE: Vertex colors that were completely white were not saved in
.wings files. (Thanks to tuxtax.) [bjorng]