Have wings_pick:do_pick/3 return what was picked
wings_pick:do_pick/3 (and wpa:pick/3) returns an updated selection,
as well as some other information, but not which object and element in
the object that was picked.

Callers that need to know which element was selected can find out
by comparing the returned selection with the previous selection, but
that is tedious and error-prone. Therefore, change the return value of
do_pick/3 to return {Id,Item,MirrorSide} instead of just MirrorSide in
the second element of the result tuple.

Also update wpc_connect_tool to handle the new return value, and
update wpc_tweak to use the new return value to fix a virtual mirror
bug.

All other callers ignore the second element in the tuple, so there
is no need to update them.

NOTE: Selecting two objects, one with a virtual mirror and one
without a virtual mirror, and Tweaking the object with the virtual
mirror would cause a crash. (Thanks to BlackHarmo.)

For developers: wings_pick:do_pick/3 and wpa:pick/3 returns more
information about what was picked in the second element of the
result tuple. Plug-ins that use the second element must
be updated. [bjorng]