On Windows, place wings crash dumps on the desktop
On Windows, we currently place the wings_crash.dump file in the
installation directory for Wings (usually in "c:/Program Files").
The problem is that on Windows Vista and later, the "Program Files"
directory is is write-protected (except for installers).

Therefore, place the crash dump on the desktop instead, if we can
find it using the registry. Failing that, we try the root of the
current drive, and if that too fails, the current directory.

NOTE: On Windows, the wings_crash.dump file will now be written to the
Desktop, since writing it to the installation directory for Wings
would not work in Windows Vista or Windows 7. [bjorng]