MacOS X/Intel: Fix reading of JPEG/PNG/GIF images
Fix a long-standing bug on Intel Macs by adding a 'native'
specifier to the BytesPerRow field. In practice, on Leopard
image reading happens to work (at least most of the time) despite
the bug, but on Snow Leopard the bug strikes every time.

NOTE: It was not possible to load JPEG, PNG, or GIF images (either
as textures or using the Image Plane commands) on Snow Leopard.
(Thanks to Riolan and nobrandname.) [bjorng]