Refresh the AutoUV window when Geom window updates UV coordinates
AutoUV will only rebuild its charts if the edge table has been
updated. Since 1.1.5, UV coordinates are not kept in the edge
table any longer, so a change of UV coordinates done by the
Geometry window (for instances, by undoing) will not force
rebuilding of the charts.

Make sure that we rebuild the charts if either the edge table
OR vertex attributes have changed.

NOTE: Undoing a change of UV coordindates would not update the
AutoUV window. (Thanks to no204.) [bjorng]