Fix crash when creating new Geometry window
In commit 4ad808c3bf720ac0aa41ce5997c2fe05a5e14ec3,
wings_toolbar:button_help/2 was made more strict in
requiring that #sh.but is a boolean.

Unfortunately, the default initialization of the #but{} record
leaves #but.sh set to 'undefined', causing a crash when a new
Geoemetry window is created.

To avoid the crash, add a default value for #but.sh so that it
never can be 'undefined'.

Also update the code in wings.erl that create the Geometry window
to send the newly created toolbar window the current status, so
that the toolbar starts out in the correct selection mode.

While at it, clean up some indentation in wings.erl.

NOTE: Creating a new Geometry window and moving the mouse cursor
over the toolbar of the new window before doing anything else would
cause Wings to crash. (Thanks to BlackHarmo.) [bjorng]