Optimize Separate when there are many sub-objects
wings_util:array_smallest_key/1 gets increasingly slower as the
number of 'undefined' entries before the first defined entry
in the array increases. It is a big win to keep track of the
next possible position where we can find a defined entry.

We can also save some time in separate/3 if we avoid putting edges
that we have already copied into the working set.