Add wings_we:fast_rebuild/1
When undoing, we only want to rebuild the 'vc' and 'fs' tables,
but not do any GC of the 'vp' table (doing a GC will not do any
good, but will probably decrease sharing with the rest of
the Undo history).

While at it, correct the comments to for rebuild/1 to note that
vertex GC always happen.